major overhaul of the server code design.
